refactor(test): move HTTPRequest.parse tests into their own file
The HTTPRequest.parse unit tests lived inside DebugRPCServerTests, whose class
body sits at the type_body_length cap (the file already carries a file_length
disable). Relocate them verbatim to a dedicated HTTPRequestTests.swift so the
parser tests sit next to the type under test and the server test class regains
headroom. Pure move; no test logic changes.
